+++ src/core/server/api/consts/constWebView.ts
@@ -2,4 +2,5 @@ import { ServerView } from '../../systems/view';
 
 export const webviewConst = {
     emit: ServerView.emit,
+    closePages: ServerView.closePages,
 };